requirements.txt
azure-ai-projects 
azure-ai-inference
azure-identity
python-dotenv
azure-storage-blob
requests
azure-ai-documentintelligence
pydantic

opentelemetry-api 
opentelemetry-sdk 
azure-monitor-opentelemetry-exporter 
opentelemetry-instrumentation-requests

applicationinsights
aiohttp

openai

agent-framework